Move FtpUploader to lib
