+++ randovania/games/dread/gui/dialog/game_export_dialog.py
@@ -12,7 +12,6 @@ from PySide6 import QtGui, QtWidgets
 
 from randovania.games.dread.exporter.game_exporter import DreadGameExportParams, DreadModPlatform
 from randovania.games.dread.exporter.options import DreadPerGameOptions
-from randovania.games.dread.gui.dialog.ftp_uploader import FtpUploader
 from randovania.games.game import RandovaniaGame
 from randovania.gui.dialog.game_export_dialog import (
     GameExportDialog,
@@ -26,6 +25,7 @@ from randovania.gui.dialog.game_export_dialog import (
 )
 from randovania.gui.generated.dread_game_export_dialog_ui import Ui_DreadGameExportDialog
 from randovania.gui.lib import common_qt_lib
+from randovania.lib.ftp_uploader import FtpUploader
 
 if TYPE_CHECKING:
     from collections.abc import Callable, Iterator
