+++ test/games/dread/gui/test_dread_game_export_dialog.py
@@ -8,11 +8,11 @@ from PySide6 import QtCore
 
 from randovania.games.dread.exporter.game_exporter import DreadGameExportParams, DreadModPlatform
 from randovania.games.dread.exporter.options import DreadPerGameOptions
-from randovania.games.dread.gui.dialog.ftp_uploader import FtpUploader
 from randovania.games.dread.gui.dialog.game_export_dialog import DreadGameExportDialog, serialize_path
 from randovania.games.dread.layout.dread_cosmetic_patches import DreadCosmeticPatches
 from randovania.games.game import RandovaniaGame
 from randovania.interface_common.options import Options
+from randovania.lib.ftp_uploader import FtpUploader
 
 
 @pytest.mark.parametrize("has_custom_path", [False, True])
